cannon: Move common state constants and utilities to core pkg
