+++ cannon/mipsevm/core/iface.go
@@ -41,10 +41,3 @@ type FPVM interface {
 	// GetDebugInfo returns debug information about the VM
 	GetDebugInfo() *DebugInfo
 }
-
-type CpuScalars struct {
-	PC     uint32 `json:"pc"`
-	NextPC uint32 `json:"nextPC"`
-	LO     uint32 `json:"lo"`
-	HI     uint32 `json:"hi"`
-}
