+++ op-challenger/game/fault/trace/alphabet/provider.go
@@ -6,10 +6,9 @@ import (
 	"fmt"
 	"math/big"
 
-	preimage "github.com/ethereum-optimism/optimism/op-preimage"
-
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm"
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm/core"
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/types"
+	preimage "github.com/ethereum-optimism/optimism/op-preimage"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
@@ -86,6 +85,6 @@ func BuildAlphabetPreimage(traceIndex *big.Int, claim *big.Int) []byte {
 
 func alphabetStateHash(state []byte) common.Hash {
 	h := crypto.Keccak256Hash(state)
-	h[0] = mipsevm.VMStatusInvalid
+	h[0] = core.VMStatusInvalid
 	return h
 }
