+++ op-challenger/game/fault/validator_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm"
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm/core"
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/types"
 	gameTypes "github.com/ethereum-optimism/optimism/op-challenger/game/types"
 	"github.com/ethereum/go-ethereum/common"
@@ -22,7 +22,7 @@ var (
 func TestValidate(t *testing.T) {
 	t.Run("ValidPrestates", func(t *testing.T) {
 		prestateHash := crypto.Keccak256(prestate)
-		prestateHash[0] = mipsevm.VMStatusUnfinished
+		prestateHash[0] = core.VMStatusUnfinished
 		player := &PrestateValidator{
 			load:     newMockPrestateLoader(false, common.BytesToHash(prestateHash)),
 			provider: newMockPrestateProvider(false, prestate),
@@ -78,7 +78,7 @@ func (m *mockPrestateProvider) AbsolutePreStateCommitment(_ context.Context) (co
 		return common.Hash{}, mockProviderError
 	}
 	hash := common.BytesToHash(crypto.Keccak256(m.prestate))
-	hash[0] = mipsevm.VMStatusUnfinished
+	hash[0] = core.VMStatusUnfinished
 	return hash, nil
 }
